+++ assets/javascripts/kitten/components/action/checkable-button/index.js
@@ -97,17 +97,20 @@ const StyledCheckableButton = styled(Button)`
   }
 
   &.k-Button--micro,
-  &.k-Button--tiny {
+  &.k-Button--tiny,
+  &.k-Button--small {
     --CheckableButton-dimension: ${pxToRem(15)};
     --CheckableButton-checkMargin: ${pxToRem(10)};
   }
 
-  &.k-Button--regular {
+  &.k-Button--regular,
+  &.k-Button--medium {
     --CheckableButton-dimension: ${pxToRem(20)};
     --CheckableButton-checkMargin: ${pxToRem(15)};
   }
 
   &.k-Button--big,
+  &.k-Button--large,
   &.k-Button--huge,
   &.k-Button--giant {
     --CheckableButton-dimension: ${pxToRem(24)};