+++ assets/javascripts/kitten/components/form/radio-button/index.js
@@ -7,6 +7,7 @@ import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
 import COLORS from '../../../constants/colors-config'
 import TYPOGRAPHY from '../../../constants/typography-config'
 import { ScreenConfig } from '../../../constants/screen-config'
+import { checkDeprecatedSizes } from '../../../helpers/utils/deprecated'
 
 const StyledRadioButton = styled.div`
   margin: ${pxToRem(10)} 0;
@@ -280,7 +281,8 @@ const StyledRadioButton = styled.div`
       }
     }
 
-    &.k-Form-RadioButton--big {
+    &.k-Form-RadioButton--big,
+    &.k-Form-RadioButton--large {
       .k-Form-RadioButton__label {
         min-height: ${pxToRem(60)};
         padding: ${pxToRem(20 - 2)} ${pxToRem(15)};
@@ -308,7 +310,8 @@ const StyledRadioButton = styled.div`
         border-width: ${pxToRem(6)};
       }
 
-      &.k-Form-RadioButton--big {
+      &.k-Form-RadioButton--big,
+      &.k-Form-RadioButton--large {
         .k-Form-RadioButton__label {
           min-height: ${pxToRem(80)};
           padding: ${pxToRem(30 - 2)} ${pxToRem(15)};
@@ -366,6 +369,9 @@ export const RadioButton = ({
   paragraphStyle,
   ...inputProps
 }) => {
+
+  checkDeprecatedSizes(size)
+  
   return (
     <StyledRadioButton
       className={classNames(
@@ -419,7 +425,7 @@ RadioButton.propTypes = {
     'Please use the Radio component instead',
   ),
   design: PropTypes.oneOf(['disc', 'check']),
-  size: PropTypes.oneOf(['small', 'regular', 'big']),
+  size: PropTypes.oneOf(['small', 'medium', 'large']),
   fontWeight: PropTypes.oneOf(['light', 'regular', 'bold']),
   paragraphStyle: PropTypes.bool,
 }
@@ -430,7 +436,7 @@ RadioButton.defaultProps = {
   error: false,
   disabled: false,
   design: 'disc',
-  size: 'regular',
+  size: 'medium',
   fontWeight: 'regular',
   paragraphStyle: false,
 }