+++ CHANGELOG.md
@@ -4,8 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-Change/Deprecation: `Information`: 
-  - Replace `tiny, regular, default, big` by `small, medium, large` prop size values.
+- Change/Deprecation: 
+  - Replace `tiny, regular, default, big` by `small, medium, large` 
+    prop size values to the following components:
+  - `Action`,
+  - `Typography`,
+  - `Form`,
+  - `Graphics`,
+  - `Information`
 
 ## [9.6.0] - 2022-04-04